refactor(SheetRegistry): enhance sheet open/close handling and animations
- Introduced local state management for sheet open/close animations.
- Updated the `handleOpenChange` function to synchronize local state with the store and delay state updates for animations.
- Simplified the `CartSheetContent` component by removing unnecessary state and effects related to drawer management.
